requirements.txt
alembic==1.5.4
asgiref==3.7.2
astroid==2.6.6
attrs==21.2.0
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.7
click==7.1.2
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.1
django-cors-headers==4.3.0
djangorestframework==3.14.0
filelock==3.12.4
flake8==3.9.2
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
fsspec==2023.10.0
gunicorn==20.1.0
httpie==3.2.2
huggingface-hub==0.17.3
idna==2.10
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==2.11.3
lazy-object-proxy==1.6.0
Mako==1.1.4
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mccabe==0.6.1
mdurl==0.1.2
multidict==6.0.4
numpy==1.26.1
packaging==20.9
parso==0.7.1
pluggy==0.13.1
psycopg2-binary==2.9.9
py==1.10.0
pycodestyle==2.7.0
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.16.1
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.19.2
python-editor==1.0.4
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.25.1
requests-toolbelt==1.0.0
rich==13.6.0
rope==0.19.0
safetensors==0.4.0
six==1.15.0
snowballstemmer==2.1.0
SQLAlchemy==1.3.23
sqlparse==0.4.2
tokenizers==0.14.1
toml==0.10.2
tqdm==4.66.1
transformers==4.34.1
typing_extensions==4.8.0
ujson==4.0.2
urllib3==1.26.4
Werkzeug==2.0.2
whitenoise==6.6.0
wonderwords==2.2.0
wrapt==1.12.1
yapf==0.31.0
openai==0.28.1

